requirements.txt
amqp==2.5.2
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.0.1
django-allauth==0.41.0
django-bootstrap-form==3.4
django-markdown-deux==1.0.5
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
Markdown==3.2
markdown2==2.3.8
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.8
vine==1.3.0
zipp==2.2.0
